fix: validation not working on performance and target endpoints add: api examples for documentation on performance and target endpoints add: retrieve employee's data for exporting on DRH

+++ database/migrations/2024_05_08_070016_create_user_performances_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('performance_id');
-            $table->tinyInteger('work_performance_score');
+            $table->float('work_performance_score', 3, 2);
             $table->timestamp('created_at');
             $table->timestamp('updated_at')->nullable();
